--- src/main/java/com/uni/cookoff/dto/RunCodeResponse.java
@@ -1,18 +1,0 @@
-package com.uni.cookoff.dto;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class RunCodeResponse {
-    private List<JudgeResponse> result;
-    private int testCasesPassed;
-}
