+++ src/main/java/com/uni/cookoff/dto/request/BatchSubmissionRequest.java
@@ -1,4 +1,4 @@
-package com.uni.cookoff.dto;
+package com.uni.cookoff.dto.request;
 
 
 import lombok.AllArgsConstructor;
@@ -6,10 +6,12 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class JudgeToken {
-    private String token;
+public class BatchSubmissionRequest {
+    private List<JudgeSubmission> submissions;
 }
