+++ src/main/java/com/uni/cookoff/dto/response/JudgeStatus.java
@@ -1,4 +1,4 @@
-package com.uni.cookoff.dto;
+package com.uni.cookoff.dto.response;
 
 
 import lombok.AllArgsConstructor;
@@ -6,12 +6,11 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class BatchSubmissionRequest {
-    private List<JudgeSubmission> submissions;
+public class JudgeStatus {
+    private String id;
+    private String description;
 }
