+++ src/main/java/com/uni/cookoff/dto/response/RunCodeResponse.java
@@ -1,4 +1,4 @@
-package com.uni.cookoff.dto;
+package com.uni.cookoff.dto.response;
 
 
 import lombok.AllArgsConstructor;
@@ -6,10 +6,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class SubmissionResponse {
-    private String submissionId;
+public class RunCodeResponse {
+    private List<JudgeResponse> result;
+    private int testCasesPassed;
 }
